requirements.txt
streamlit
langchain
langgraph
google-generativeai
pymongo
python-dotenv
pandas
numpy
geopy
twilio
streamlit-aggrid
plotly
seaborn
scikit-learn
langchain-google-genai
pydeck